refactor: adapter 제거 및 StoreCategoryService 생성
